requirements.txt
numpy
opencv-python-headless
pillow
pytesseract
